refactor: cleanup the directory structure
