+++ pkg/notationfactory/client.go
@@ -1,13 +1,9 @@
 package notationfactory
 
 import (
-	"encoding/json"
-	"fmt"
-	"io/fs"
 	"os"
 	"path/filepath"
 	"strings"
-	"sync"
 
 	"github.com/nirmata/kyverno-notation-verifier/kubenotation/utils"
 	"github.com/nirmata/kyverno-notation-verifier/types"
@@ -15,16 +11,17 @@ import (
 	"github.com/notaryproject/notation-go/dir"
 	"github.com/notaryproject/notation-go/plugin"
 	"github.com/notaryproject/notation-go/verifier"
-	"github.com/notaryproject/notation-go/verifier/trustpolicy"
 	"github.com/notaryproject/notation-go/verifier/truststore"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
 
-type notationverifierfactory struct {
-	verifiers map[string]*notation.Verifier
-	log       *zap.SugaredLogger
-	lock      sync.Mutex
+type NotationVeriferFactory interface {
+	// RefreshVerifiers will remove all the existing verifiers and create new ones using the trust policies in notation directory
+	RefreshVerifiers() error
+
+	// GetVerifier returns a verifier based on the trust policy in request or the default trustpolicy in trust policy env
+	GetVerifier(requestData *types.RequestData) (*notation.Verifier, error)
 }
 
 func NewNotationVerifierFactory(logger *zap.SugaredLogger) NotationVeriferFactory {
@@ -32,11 +29,6 @@ func NewNotationVerifierFactory(logger *zap.SugaredLogger) NotationVeriferFactor
 		verifiers: make(map[string]*notation.Verifier),
 		log:       logger,
 	}
-}
-
-func (f *notationverifierfactory) clear() {
-	f.verifiers = make(map[string]*notation.Verifier)
-	f.log.Info("Notation verifier factory cleared")
 }
 
 func (f *notationverifierfactory) RefreshVerifiers() error {
@@ -110,36 +102,4 @@ func (f *notationverifierfactory) GetVerifier(requestData *types.RequestData) (*
 	f.log.Infof("Found notation verifer for trust policy %s", trustPolicy)
 
 	return verifier, nil
-}
-
-func (f *notationverifierfactory) loadTrustPolicy(path string) (*trustpolicy.Document, error) {
-	f.log.Infof("Loading trust policy path=%s", path)
-	fileInfo, err := os.Lstat(path)
-	if err != nil {
-		if errors.Is(err, os.ErrNotExist) {
-			return nil, fmt.Errorf("trust policy is not present %s", path)
-		}
-		return nil, err
-	}
-
-	mode := fileInfo.Mode()
-	if mode.IsDir() || mode&fs.ModeSymlink != 0 {
-		return nil, fmt.Errorf("trust policy is not a regular file (symlinks are not supported) path: %s", path)
-	}
-
-	jsonFile, err := os.Open(path)
-	if err != nil {
-		if errors.Is(err, os.ErrPermission) {
-			return nil, fmt.Errorf("unable to read trust policy due to file permissions, please verify the permissions of %s", path)
-		}
-		return nil, err
-	}
-	defer jsonFile.Close()
-
-	policyDocument := &trustpolicy.Document{}
-	err = json.NewDecoder(jsonFile).Decode(policyDocument)
-	if err != nil {
-		return nil, fmt.Errorf("malformed trust policy path: %s", path)
-	}
-	return policyDocument, nil
 }
