+++ verifier/internal/utils.go
@@ -1,11 +1,11 @@
-package verifier
+package internal
 
 import (
 	"github.com/nirmata/kyverno-notation-verifier/types"
 	"github.com/pkg/errors"
 )
 
-func validateRequestData(req *types.RequestData) error {
+func ValidateRequestData(req *types.RequestData) error {
 	if len(req.Images.Containers) == 0 &&
 		len(req.Images.InitContainers) == 0 &&
 		len(req.Images.EphemeralContainers) == 0 {
